Refactor: 중복된 로직을 useTabNav 훅으로 분리
